DataObjects/Core_watch.php

+++ DataObjects/Core_watch.php
@@ -232,15 +232,18 @@ class Pman_Core_DataObjects_Core_watch extends DB_DataObject
 
     function onUpdate($old, $request,$roo, $event)
     {
-        // delete pending notifications
-        DB_DataObject::factory('core_notify')->query(
-            "DELETE FROM
-                core_notify
-            WHERE
-                watch_id = {$this->id}
-            AND
-                event_id < 1
-        ");
+        // becomes inactive
+        if($old->active != $this->active && !$this->active) {
+            // delete pending notifications
+            DB_DataObject::factory('core_notify')->query(
+                "DELETE FROM
+                    core_notify
+                WHERE
+                    watch_id = {$this->id}
+                AND
+                    event_id < 1
+            ");
+        }
     }
     
     function toRooSingleArray($au,$q)
